Organizes section components into seperate folder
